Telegram.Bot: update to 16.0.2

+++ CaptchaBot/Services/IUsersStore.cs
@@ -15,7 +15,7 @@ namespace CaptchaBot.Services
 
         IReadOnlyCollection<NewUser> GetAll();
 
-        NewUser Get(long chatId, int userId);
+        NewUser Get(long chatId, long userId);
         
         void Remove(NewUser user);
     }
